Date parser as separate entity
